+++ MiniMonClient/src/trace/format.cpp
@@ -83,7 +83,7 @@ namespace {
     constexpr int PTR_WIDTH = static_cast<int>(sizeof(protocol::ObjectId) * 2u);
 
     std::wstring RenderTime(int64_t kernelTime) {
-        const ULONGLONG ticks = static_cast<ULONGLONG>(kernelTime);
+        const uint64_t ticks = static_cast<uint64_t>(kernelTime);
 
         FILETIME fileTime{};
         fileTime.dwLowDateTime = static_cast<DWORD>(ticks);
@@ -98,7 +98,7 @@ namespace {
         if (!FileTimeToSystemTime(&localTime, &sysTime)) return L"TIME ERROR";
 
         // 100 ns fraction within the second
-        const ULONG subSecond = static_cast<ULONG>(ticks % 10000000ull);
+        const uint32_t subSecond = static_cast<uint32_t>(ticks % 10000000ull);
 
         return std::format(L"{:02}:{:02}:{:02}.{:07}", sysTime.wHour, sysTime.wMinute, sysTime.wSecond, subSecond);
     }
@@ -113,13 +113,13 @@ namespace {
 
 
     std::wstring RenderStackTrace(const protocol::RecordData& data) {
-        const ULONG count = data.stackFrameCount < protocol::STACK_TRACE_FRAMES ? data.stackFrameCount : protocol::STACK_TRACE_FRAMES;
+        const uint32_t count = data.stackFrameCount < protocol::STACK_TRACE_FRAMES ? data.stackFrameCount : protocol::STACK_TRACE_FRAMES;
 
         if (count == 0u) return L"";
 
         std::wstring result;
 
-        for (ULONG i = 0u; i < count; i++) {
+        for (uint32_t i = 0u; i < count; i++) {
 
             if (i > 0u) result.push_back(L'|');
 
@@ -215,7 +215,7 @@ namespace mimo {
                     columns[MAJOR]         = names::RenderMajorFunction(data.callbackMajorId);
                     columns[MINOR]         = names::RenderMinorFunction(data.callbackMajorId, data.callbackMinorId);
                     columns[NAME]          = EscapeCsvField(text::MarkTruncated(text::Extract(data.name), data.truncated & protocol::TRUNCATED_NAME));
-                    columns[STATUS]        = std::format(L"{:08X}", static_cast<ULONG>(data.status));
+                    columns[STATUS]        = std::format(L"{:08X}", static_cast<uint32_t>(data.status));
                     columns[INFORMATION]   = std::format(L"{:0{}X}", data.information, PTR_WIDTH);
                     columns[DETAILS]       = EscapeCsvField(details::Render(data));
                     columns[IRP_FLAGS]     = std::format(L"{:08X}", data.irpFlags);
@@ -225,7 +225,7 @@ namespace mimo {
                     columns[ARG3]          = std::format(L"{:0{}X}", data.parameters.others.argument3, PTR_WIDTH);
                     columns[ARG4]          = std::format(L"{:0{}X}", data.parameters.others.argument4, PTR_WIDTH);
                     columns[ARG5]          = std::format(L"{:0{}X}", data.parameters.others.argument5, PTR_WIDTH);
-                    columns[ARG6]          = std::format(L"{:0{}X}", static_cast<ULONGLONG>(data.parameters.others.argument6), PTR_WIDTH);
+                    columns[ARG6]          = std::format(L"{:0{}X}", static_cast<uint64_t>(data.parameters.others.argument6), PTR_WIDTH);
                     columns[REPARSE_TAG]   = names::RenderReparseTag(data.reparseTag);
                 }
